+++ agent-allure-cucumber6/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 ## [5.2.2]
+### Changed
+- Agent version updated on [5.3.5](https://github.com/reportportal/agent-java-cucumber6/releases/tag/5.3.5), by @HardNorth
 
 ## [5.2.1]
 ### Changed